Create eight-of-the-biggest-lies-in-APIs
Submitting an article for the APIs You Won't Hate blog. Will need a cover image. In addition, @philsturgeon thought it might be worthwhile to springboard from this piece and link to other, existing articles (for example, could see the cache lie easily linking to the published work on caching). Somebody with knowledge of the deep cuts will need to add those, if desired.